# server/openapi/atlas-v1.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Atlas PPM API",
    "description": "Portfolio & Project Management REST API (/api/v1).",
    "version": "v1"
  },
  "paths": {
...
    "/api/v1/programs/{id}": {
      "patch": {
        "tags": [
          "Atlas.Api"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProgramReq"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "delete": {
        "tags": [
          "Atlas.Api"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/api/v1/projects/{id}/capacity": {
      "get": {
        "tags": [
          "Atlas.Api"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CapacityDto"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CapacityDto": {
        "type": "object",
        "properties": {
          "assigned": {
            "type": "integer",
            "format": "int32"
          },
          "overCount": {
            "type": "integer",
            "format": "int32"
          },
          "highOps": {
            "type": "integer",
            "format": "int32"
          },
          "people": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CapacityRowDto"
            },
            "nullable": true
          },
          "unknown": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "CapacityRowDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "nullable": true
          },
          "role": {
            "type": "string",
            "nullable": true
          },
          "initials": {
            "type": "string",
            "nullable": true
          },
          "color": {
            "type": "string",
            "nullable": true
          },
          "opsPct": {
            "type": "integer",
            "format": "int32"
          },
          "projectPct": {
            "type": "integer",
            "format": "int32"
          },
          "productPct": {
            "type": "integer",
            "format": "int32"
          },
          "util": {
            "type": "integer",
            "format": "int32"
          },
          "over": {
            "type": "boolean"
          },
          "highOps": {
            "type": "boolean"
          }
        },
        "additionalProperties": false
      },
...
      "UpdateProgramReq": {
        "type": "object",
        "properties": {
          "owner": {
            "type": "string",
            "nullable": true
          },
          "dept": {
            "type": "string",
            "nullable": true
          },
          "goal": {
            "type": "string",
            "nullable": true
          },
          "startDate": {
            "type": "string",
            "nullable": true
          },
          "endDate": {
            "type": "string",
            "nullable": true
          },
          "projects": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
    }
  }
}